requirements.txt
aiohttp==3.6.2
allure-pytest==2.8.12
allure-python-commons==2.8.12
altgraph==0.17
Appium-Python-Client==0.49
asgiref==3.2.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
BeautifulReport==0.1.2
behave==1.2.6
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
ConfigArgParse==0.15.1
crypto==1.4.1
cryptography==2.9.2
DBUtils==1.3
dill==0.3.1.1
Django==3.0.2
django-request==1.5.5
dnspython==1.16.0
Flask==1.1.1
framework==0.1.0
future==0.18.2
get==2019.4.13
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.5.19
httpcore==0.9.0
httpx==0.13.1
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
locust==0.0
locustio==0.13.5
lxml==4.5.0
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==8.1.0
msgpack==1.0.0
msgpack-python==0.5.6
multidict==4.7.5
mysql-connector-python==8.0.19
Naked==0.1.31
packaging==20.0
paramiko==2.7.1
parse==1.15.0
parse-type==0.5.2
pbr==5.4.4
pefile==2019.4.18
pluggy==0.13.1
post==2019.4.13
protobuf==3.6.1
psutil==5.7.0
public==2019.4.13
py==1.8.1
pycparser==2.19
pycryptodome==3.9.7
PyInstaller==3.6
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.6
pytest==5.4.3
pytest-dependency==0.5.1
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
pyws==1.1.5
pywss==0.0.11
PyYAML==5.3
pyzmq==18.1.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
rfc3986==1.4.0
route==0.1.4
selenium==3.141.0
Send2Trash==1.5.0
shellescape==3.8.1
six==1.13.0
sniffio==1.1.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
stevedore==1.31.0
threadpool==1.3.2
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
wcwidth==0.1.8
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
yarl==1.4.2
